Noticket: fix PMD warnings
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@16865 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ src/java/org/ejbca/core/model/ra/raadmin/EndEntityProfile.java
@@ -687,7 +687,7 @@ public class EndEntityProfile extends UpgradeableDataHashMap implements Serializ
     
     /** @return strength in bits = log2(possible chars) * number of chars rounded down */
     private int getPasswordStrength(int numerOfDifferentChars, int passwordLength) {
-    	return Double.valueOf(Math.floor(Math.log((double)numerOfDifferentChars)/Math.log(2)) * (double)passwordLength).intValue();
+    	return (int) (Math.floor(Math.log((double)numerOfDifferentChars)/Math.log(2)) * (double)passwordLength);
     }
 
     /** @return the minimum strength that a password is allowed to have in bits */
